+++ crates/supervisor/core/src/event/chain.rs
@@ -33,10 +33,4 @@ pub enum ChainEvent {
         /// hash.
         replacement: BlockReplacement,
     },
-
-    /// A L1 finalized event, indicating that a new L1 block has been finalized.
-    L1Finalized {
-        /// The [`BlockInfo`] of the finalized L1 block.
-        block: BlockInfo,
-    },
 }
