+++ crates/supervisor/storage/src/chaindb.rs
@@ -29,16 +29,13 @@ pub struct ChainDb {
     /// Current L1 block reference, used for tracking the latest L1 block processed.
     /// In-memory only, not persisted.
     current_l1: RwLock<Option<BlockInfo>>,
-    /// Finalized L1 block reference, used for tracking the finalized L1 block.
-    /// In-memory only, not persisted.
-    finalized_l1: RwLock<Option<BlockInfo>>,
 }
 
 impl ChainDb {
     /// Creates or opens a database environment at the given path.
     pub fn new(path: &Path) -> Result<Self, StorageError> {
         let env = init_db_for::<_, crate::models::Tables>(path, DatabaseArguments::default())?;
-        Ok(Self { env, current_l1: RwLock::new(None), finalized_l1: RwLock::new(None) })
+        Ok(Self { env, current_l1: RwLock::new(None) })
     }
 
     /// initialises the database with a given anchor derived block pair.
@@ -153,14 +150,6 @@ impl HeadRefStorageReader for ChainDb {
         guard.as_ref().cloned().ok_or(StorageError::FutureData)
     }
 
-    fn get_finalized_l1(&self) -> Result<BlockInfo, StorageError> {
-        let guard = self.finalized_l1.read().map_err(|err| {
-            error!(target: "supervisor_storage", %err, "Failed to acquire read lock on finalized_l1");
-            StorageError::LockPoisoned
-        })?;
-        guard.as_ref().cloned().ok_or(StorageError::FutureData)
-    }
-
     fn get_safety_head_ref(&self, safety_level: SafetyLevel) -> Result<BlockInfo, StorageError> {
         self.env.view(|tx| SafetyHeadRefProvider::new(tx).get_safety_head_ref(safety_level))?
     }
@@ -188,33 +177,6 @@ impl HeadRefStorageWriter for ChainDb {
             }
         }
         *guard = Some(block);
-        Ok(())
-    }
-
-    fn update_finalized_l1(&self, block: BlockInfo) -> Result<(), StorageError> {
-        let mut guard = self
-            .finalized_l1
-            .write()
-            .map_err(|err| {
-                error!(target: "supervisor_storage", %err, "Failed to acquire write lock on finalized_l1");
-                StorageError::LockPoisoned
-            })?;
-
-        // Check if the new block number is greater than the current finalized block
-        if let Some(ref current) = *guard {
-            if block.number <= current.number {
-                error!(target: "supervisor_storage",
-                    current_block_number = current.number,
-                    new_block_number = block.number,
-                    "New finalized block number is not greater than current finalized block number",
-                );
-                return Err(StorageError::BlockOutOfOrder);
-            }
-        }
-        *guard = Some(block);
-
-        // todo: update safety head ref for finalized level
-
         Ok(())
     }
 
@@ -544,35 +506,6 @@ mod tests {
         // Update with a lower block number should error
         let block3 = BlockInfo { number: 15, ..Default::default() };
         let err = db.update_current_l1(block3).unwrap_err();
-        assert!(matches!(err, StorageError::BlockOutOfOrder));
-    }
-
-    #[test]
-    fn test_update_and_get_finalized_l1() {
-        let tmp_dir = tempfile::TempDir::new().unwrap();
-        let db_path = tmp_dir.path().join("chaindb_finalized_l1");
-        let db = ChainDb::new(&db_path).unwrap();
-
-        let block1 = BlockInfo { number: 100, ..Default::default() };
-        let block2 = BlockInfo { number: 200, ..Default::default() };
-
-        // Initially, get_finalized_l1 should return FutureData error
-        let err = db.get_finalized_l1().unwrap_err();
-        assert!(matches!(err, StorageError::FutureData));
-
-        // Update finalized_l1 with block1
-        db.update_finalized_l1(block1).unwrap();
-        let got = db.get_finalized_l1().unwrap();
-        assert_eq!(got, block1);
-
-        // Update with a higher block number
-        db.update_finalized_l1(block2).unwrap();
-        let got = db.get_finalized_l1().unwrap();
-        assert_eq!(got, block2);
-
-        // Update with a lower block number should error
-        let block3 = BlockInfo { number: 150, ..Default::default() };
-        let err = db.update_finalized_l1(block3).unwrap_err();
         assert!(matches!(err, StorageError::BlockOutOfOrder));
     }
 }
