+++ crates/supervisor/storage/src/traits.rs
@@ -159,13 +159,6 @@ pub trait HeadRefStorageReader: Debug {
     /// * `Err(StorageError)` if there is an issue retrieving the reference.
     fn get_current_l1(&self) -> Result<BlockInfo, StorageError>;
 
-    /// Retrieves the finalized L1 block reference from the storage.
-    ///
-    /// # Returns
-    /// * `Ok(BlockInfo)` containing the finalized L1 block reference.
-    /// * `Err(StorageError)` if there is an issue retrieving the reference.
-    fn get_finalized_l1(&self) -> Result<BlockInfo, StorageError>;
-
     /// Retrieves the current [`BlockInfo`] for a given [`SafetyLevel`].
     ///
     /// # Arguments
@@ -195,16 +188,6 @@ pub trait HeadRefStorageWriter: Debug {
     /// * `Err(StorageError)` if there is an issue updating the reference.
     fn update_current_l1(&self, block: BlockInfo) -> Result<(), StorageError>;
 
-    /// Updates the finalized L1 block reference in the storage.
-    ///
-    /// # Arguments
-    /// * `block` - The new [`BlockInfo`] to set as the finalized L1 block reference.
-    ///
-    /// # Returns
-    /// * `Ok(())` if the reference was successfully updated.
-    /// * `Err(StorageError)` if there is an issue updating the reference.
-    fn update_finalized_l1(&self, block: BlockInfo) -> Result<(), StorageError>;
-
     /// Updates the safety head reference for a given [`SafetyLevel`].
     ///
     /// # Arguments
@@ -228,3 +211,25 @@ pub trait HeadRefStorageWriter: Debug {
 pub trait HeadRefStorage: HeadRefStorageReader + HeadRefStorageWriter {}
 
 impl<T: HeadRefStorageReader + HeadRefStorageWriter> HeadRefStorage for T {}
+
+/// Provides an interface for managing the finalized L1 block reference in the storage.
+/// 
+/// This trait defines methods to update and retrieve the finalized L1 block reference.
+pub trait FinalizedL1Storage {
+    /// Updates the finalized L1 block reference in the storage.
+    ///
+    /// # Arguments
+    /// * `block` - The new [`BlockInfo`] to set as the finalized L1 block reference.
+    ///
+    /// # Returns
+    /// * `Ok(())` if the reference was successfully updated.
+    /// * `Err(StorageError)` if there is an issue updating the reference.
+    fn update_finalized_l1(&self, block: BlockInfo) -> Result<(), StorageError>;
+
+    /// Retrieves the finalized L1 block reference from the storage.
+    ///
+    /// # Returns
+    /// * `Ok(BlockInfo)` containing the finalized L1 block reference.
+    /// * `Err(StorageError)` if there is an issue retrieving the reference.
+    fn get_finalized_l1(&self) -> Result<BlockInfo, StorageError>;
+}
